move hook to correct folder
